separate Desktop register & Mobile register
