+++ src/component/register/MobileRegister.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const MobileRegister = () => {
+  return (
+    <div>MobileRegister</div>
+  )
+}
+
+export default MobileRegister
